% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_high_daily}
\alias{asset_high_daily}
\title{high level asset returns}
\format{An xts
\describle{
\item {large.cap.value}{large capital growth corporate, code=399373}
\item {large.cap.growth}{large capital growth corporate, code=399382}
\item {small.cap.value}{small capital value corporate, code=399377}
\item {small.cap.growth}{small capital growth corporate, code=399376}
\item {middle.cap.value}{middle capital value corportate, code=399375}
\item {midle.cap.growth}{middle capital corporate with growth opportunities, code=399374}
\item {commodity.stock}{cmmodity index in stock market, code=000979}
\item {commodity.futures}{commodity index in futures market, code=H30009}
\item {convertiable}{convertable bond code=000832}
\item {bond.5yr}{credit bond with maturity almost equals to 5 years,code=000101}
\item {szAApup}{credit bond with grade geq AA+,code=399298}
\item {szAApdown}{credit bond with grade less AA+, code=399299}
\item {treasure}{teasure bond index code=000012}
}}
\usage{
asset_high_daily
}
\description{
high level asset returns
}
\keyword{datasets}
