% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_high_monthly}
\alias{asset_high_monthly}
\title{monthly return of high assets}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 52 rows and 13 columns.}
\usage{
asset_high_monthly
}
\description{
monthly return of high assets
}
\seealso{
asset_high_daily
}
\keyword{datasets}
