% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_top_daily}
\alias{asset_top_daily}
\title{top level asset returns
data are download from finance.sina.com.cn through tushare}
\format{An xts
\describle{
\item{stock}{hs300, index of 300 large cap stock in Shanghai and Shenzhen stock exchange market, code ='000300'}
\item{treasure}{index of treasure bonds in the exchange market. code ='000012'}
\item{corporate.bond}{index of corporate bonds in the exchange market. code='000013'}
\item{convertable}{index of convertable bond, code='000832'}
\item{commodity}{index of commodity futures code='H30009'}
}}
\usage{
asset_top_daily
}
\description{
top level asset returns
data are download from finance.sina.com.cn through tushare
}
\keyword{datasets}
