% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_top_monthly}
\alias{asset_top_monthly}
\title{monthly return of top assets}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 99 rows and 5 columns.}
\usage{
asset_top_monthly
}
\description{
monthly return of top assets
}
\seealso{
asset_top_daily
}
\keyword{datasets}
