% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_top_quarterly}
\alias{asset_top_quarterly}
\title{quarterly return of top assets}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 34 rows and 5 columns.}
\usage{
asset_top_quarterly
}
\description{
quarterly return of top assets
}
\seealso{
asset_top_daily
}
\keyword{datasets}
