% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_top_weekly}
\alias{asset_top_weekly}
\title{weekly return of top assets}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 423 rows and 5 columns.}
\usage{
asset_top_weekly
}
\description{
weekly return of top assets
}
\seealso{
asset_top_daily
}
\keyword{datasets}
