% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_all}
\alias{generate_all}
\title{Function to generate all return data}
\usage{
generate_all()
}
\description{
currently, top level assets return, high level assets return of general asset classes (stock,bond,commodity,cash,real estate). returns of representitive indexes of stock market, second level industrial indexes of stock (provided by ShenWan), randomly picked stocks are computed and stored under the directory of data. Only called when you want to update the data from database
}
\examples{
init()
genrate_all()
}
