% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{getReturnFromDB}
\alias{getReturnFromDB}
\title{get returns from my postgresql database}
\usage{
getReturnFromDB(assets, from, to = NULL, frequency = "daily",
  table.name = "index_ohlc_day", asset.names = NULL,
  join.method = "inner")
}
\arguments{
\item{assets}{list of asset codes}

\item{from}{the starting date in string format}

\item{to}{the ending date in string format, no limit if this is set to NULL (default)}

\item{frequency}{can only take values from 'daily','weekly','monthly' and 'quarterly'}

\item{table.name}{tablename in postgresql database. We will try to retrieve price information from the table}

\item{asset.names}{a readable name for each asset, in a list format}

\item{join.method}{take values from "inner", "outer", "left inner" and "right inner", but may not support by db driver.}
}
\description{
do not call it if you do not have the same dataabse with you, this package is designed to supply data only.
}
\examples{
getReturnFromDB(c('000300','000012'),'2000-01-01',to=NULL,table.name='index_ohlc_day',frequency='daily',asset.names=c('hs300','teasure.bond'))


}
