% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{randomstocks}
\alias{randomstocks}
\title{randomstocks: sampling random number of stock within each industry}
\usage{
randomstocks(n, older.than = "20000101")
}
\arguments{
\item{n}{how many stock sampled from each industry}

\item{older.than}{a string of date in the format '20000101' only, which is a creteria for selected stocks whose "time to Market" must be older than it.}
}
\description{
randomly sampled stock code and names are returned and save as dataset in the data directory
}
\examples{
randomstocks(3,older.than='20000203')

}
