% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{saveReturns}
\alias{saveReturns}
\title{saveReturns: retrieve data from related data table and save them into a rda file}
\usage{
saveReturns(assets, asset.names, data.name, where, from, to = NULL)
}
\arguments{
\item{assets}{asset codes ,e.g. '000300' is the code for index hs300}

\item{asset.names}{readable names for each asset}

\item{data.name}{the R variable name you want when load the data into memory. e.g. "stocks"}

\item{where}{the asset class of assets. now only "stock" and "index" are supported. "Futures", "funds" will be supported soon.}

\item{from}{the starting date of asset prices}

\item{to}{the ending date of asset prices}
}
\examples{
st=randomstocks(3)
saveReturns(st[,1],st[,2],'stocks',where='stock',from='2000-01-01')
}
