% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_index_quarterly}
\alias{stock_index_quarterly}
\title{stock indexes quarter return}
\format{an xts
\describle{
\item {large.cap.value}{large capital growth corporate, code=399373}
\item {large.cap.growth}{large capital growth corporate, code=399382}
\item {small.cap.value}{small capital value corporate, code=399377}
\item {small.cap.growth}{small capital growth corporate, code=399376}
\item {middle.cap.value}{middle capital value corportate, code=399375}
\item {midle.cap.growth}{middle capital corporate with growth opportunities, code=399374}
\item{stock}{hs300, index of 300 large cap stock in Shanghai and Shenzhen stock exchange market, code ='000300'}
\item{middle.small.cap}{index of middle or small corporates, code=399101}
\item{growth.enterprise}{index of growth.enterprise, code=399606}
}}
\usage{
stock_index_quarterly
}
\description{
stock indexes quarter return
}
\keyword{datasets}
