% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_daily}
\alias{stocks_daily}
\title{individual stock daily returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.}
\format{an xts}
\usage{
stocks_daily
}
\description{
individual stock daily returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.
}
\keyword{datasets}
