% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_monthly}
\alias{stocks_monthly}
\title{individual stock monthly returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.}
\format{an xts}
\usage{
stocks_monthly
}
\description{
individual stock monthly returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.
}
\keyword{datasets}
