% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_quarterly}
\alias{stocks_quarterly}
\title{individual stock quarter returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.}
\format{an xts}
\usage{
stocks_quarterly
}
\description{
individual stock quarter returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.
}
\keyword{datasets}
