% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_weekly}
\alias{stocks_weekly}
\title{individual stock weekly returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.}
\format{an xts}
\usage{
stocks_weekly
}
\description{
individual stock weekly returns. stocks are randomly sampled from each industries. At most 3 stock within each industry.
}
\keyword{datasets}
